voice_python_mentor/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
replicate>=0.11.0
playsound3>=1.0.0
